backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
tiktoken>=0.5.0
openai>=1.0.0
resend>=2.13.0
